rozdzielenie na pliki
