+++ my_class/Image/plansza_marker_pionki.png
